+++ changelog.md
@@ -3,7 +3,8 @@
 ## [0.0.7] - 2023-04-02
 ### Added
 - counting and displaying deltaTime and FramesPerSecond
-
+- MovingObject class
+- MovingObject has velocity in all axes, its position changes depending on deltaTime 
 
 ## [0.0.6] - 2023-03-31